Bug#36313072: mysqld crash in Item_ref::real_item
While resolving an expression in QUALIFY clause, if the expression is
resolved against a field in a table or a merged view, it is added to
the select expression list as a hidden item. In the case when
the resolver later reduces the QUALIFY clause to an always false or an
always true condition, it removes the expressions present in the
QUALIFY clause and replaces them with a TRUE or FALSE condition.
Care is taken to not remove these expressions when there
are still references to these objects. However, for the case where the
hidden item is added to the select expression list, reference count is
not incremented correctly resulting in removal of the expressions.
This leads to problems later when the added hidden item is looked at.
Solution is to increment the reference count when the hidden item is
added to the select expression list. We already increment the ref count
once, which accounts for the reference from Item_ref object we are
trying to resolve. Now we account for the presence in the select
expression list.

Change-Id: Ife3e51f1f38fe030ab2381c2ef7354d033dff5bf