Bug#36357380 testestNdbApi -n Bug28443 leaves Error 9003 Inserted after running
Error is injected to TRPMAN Proxy and then proxy resends the error
to all workers where it is eventually cleared. But, in TRPMAN proxy
instance error is never cleared.
Fix:
Clear error in test side (Insert Error 0) to force it to be cleared
in all TRPMAN instances (including TRPMAN proxy).

Change-Id: I7a024dca555c8c684e589c62ea2963722b7531a7