Bug#33390949: Execute a statement with secondary engine, two entries in performance schema [1/2, regular statements, noclose]
When executing a query which involves tables in a secondary engine,
there could be multiple entries for the statement in performance
schema, even though the query was only executed once.

This happened because check_secondary_engine_statement() reported to
performance schema that the current statement ended and a new
statement started when it was decided to retry the statement in a
different storage engine.

Fixed by removing that code from check_secondary_engine_statement(),
so that only a single instance is recorded in performance schema.

The test case for the bug was added to the
secondary_engine.query_preparation test, which had previously been
disabled wholesale awaiting the fix for bug#36306426. This patch
re-enables the test, and instead disables the few parts of the test
that need bug#36306426 to be fixed.

Change-Id: Ib7789816bb16d3211396e4f7c7865896c7352c67