Bug#36338562 PluginConfig::get_default forces std::string creation
The PluginConfig class is the glue between between the Config system and
the plugin's configuration.

It's methods:

- get_defaults()
- is_required()

lookup config-option against the internal names. They currently use
'const std::string &' which means the caller has to instantiate a
std::string even if the input is

- std::string_view
- const char *

Better is to use a std::string_view.

Change
======

- use std::string_view instead of 'const std::string &' for
  string-lookup in get_defaults() and is_required()
- change internal maps from std::map<std::string, std::string> to
  std::map<std::string_view, std::string> to reduce memory-copies.

Change-Id: Ib569d98c58171294aa24248c478946ff89e543af