Bug#35851009: rpl_innodb_bug28430 fails with WRITESET, ROW and GTID_MODE=OFF
* Issue

This test uses --innodb_autoinc_lock_mode=0.

This in itself is not an issue, but when applier concurrency is high
in the replica (due to the usage of WRITESET), it does lead to
deadlocks being reported among workers for the auto increment lock.

At the same time the replica-preserve-commit-order option causes the
engine to signal transactions as being a cause of a deadlock, and to
allow the workers that have transactions ahead on the queue that need
the lock, these signaled transactions would rollback to allow the
overall progress of the applier.

The issue with these autoinc deadlocks is that they are ephemeral,
i.e., the transaction that declares the deadlock eventually gets the
lock in question and is able to progress. On the other hand, the
transactions marked to be rolled back sometimes reach commit and all
previous transactions they would wait on already committed, so it
bypasses the commit order manager wait, and also the rollback process.

The result of the above events is that a transaction marked for
rollback does not rollback, but after commit the worker still sees it
as deadlock causing transaction, assumes it was rolled back and tries
to retry it. The worker when retrying an already applied transaction
then fails with a duplicated key error.

If GTIDs are enabled, then the transaction is skipped when retried.

* Solution

If a transaction reaches commit, if in the commit order it doesn't
need to wait, then the deadlock flag it might have is set to false.

Change-Id: Ic1f2a948d8efe0e04b0fd66d6034ee26a6053e8e