Bug#36332697 IN predicate containing EXCEPT ALL set operation gives wrong result
This error happens both with hash_set_operations on or off. If we look
    at the explain, we can see why it happens:

    -> Filter: <in_optimizer>(c.pk,<exists>(select #2))
        -> Covering index scan on c using idx_c_col_datetime_key
        -> Select #2 (subquery in condition; dependent)
            -> Limit: 1 row(s)  <--------------------- OK optimization
                -> Table scan on <except temporary>
                    -> Except all materialize
                        -> Limit: 1 row(s)  <--------------------------- problem
                            -> Table scan on <union temporary>
                                -> Union all materialize
                                    -> Filter: (<cache>(c.pk) = <ref_null_helper>(2))
                                        -> Rows fetched before execution
                                    -> Filter: (<cache>(c.pk) = <ref_null_helper>(2))
                                        -> Rows fetched before execution
                        -> Filter: (<cache>(c.pk) = <ref_null_helper>(2))
                            -> Rows fetched before execution

There is a limit node on top of the left EXCEPT operand right over the
UNION ALL node which shouldn't be there. It used to be a clever
optimization when MySQL only had UNION set operations, but it clearly
is wrong for EXCEPT ALL. For EXCEPT DISTINCT, UNION and INTERSECT, it
is fine, though.

The solution is to skip pushed down limits in the presence of EXCEPT
ALL inside subqueries, unless it is the top query block in the
subquery's query_expression, i.e. we retain the top-most limit 1 above.

Change-Id: Idf784cbfbe8efbaca03ad17c8c42d73ab7acaa1a