Bug#36356479 decouple TRPMAN error insert handling from CMVMI
Problem:
Currently there is a mix between CMVMI and TRPMAN error insert handling.
Part of error insert values handled in TRPMAN were originally
handled in CMVMI but, when TRPMAN was introduced and all transporter
stuff was moved to TRPMAN, they were moved from CMVMI to TRPMAN as well.
Also, as part of the moving of transporter handling to TRPMAN, part of
the CMVMI error code range was reserved to TRPMAN [9500, 9899] causing
CMVMI to have, in fact, two different error range [9000, 9499]
and [9900, 9999].

This mix of error insert handling in both CMVMI and TRPMAN is
undesirable because in some cases signals are, unnecessarily, sent to
CMVMI and then, re-sent to TRPMAN and, in addition, it can cause
difficulties reading error insert code.

Solution:
- TRPMAN error insert handling decoupled from CMVMI.
- Error range for CMVMI and TRPMAN redefined as:
   - TRPMAN [9000, 9599]
   - CMVMI  [9600, 9999]
- All existing error insert code preserved with same behaviour as before.
- Existing error in the [9000, 9599] range, handled in CMVMI, moved
to TRPMAN.
- Handling of dump 9006 moved to TRPMAN as well.

This patch also fixes testNodeRestart -n Bug24717.
Test dumps code 9002 to data node. TRPMAN sets error 9002 in self
block, but it never hit because handling of error 9002 is done in
CMVMI instead TRPMAN. (It was probably missed in the migration of
transporter stuff from CMVMI to TRPMAN).
Fix:
execNODE_START_REP Implemented in TRPMAN block so error 9002 can be
handled as expected.
Also, TRPMAN block added to all_blocks list in NDBCNTR in order
to make NODE_START_REP to be sent to TRPMAN instances.

Change-Id: Ibf56def74818f5043b43c40e39ecdaa1e4e02974