WL#15699 Reject foreign key references to non unique key and partial key
Followup patch: fk_on_std_key_upgrade.test created as part of
WL#15699 checks upgrade from 8.2. This test is applicable only
for the 8.4 release and not for trunk(9.0) because an upgrade is
allowed only to immediate next LTS (i.e 8.4 in this case).
Therefore, removing this test and the associated database zip
file from trunk.

Change-Id: I10790ac269a926de7ac859c54ef6b8232ad35476