Bug#36332645 InnoDB recovery shouldn't make 0.5s pauses when waiting for deltas to be applied
Instead of using a loop to check if number of pages waiting for changes
to be applied drops to zero every 0.5s, simply define this event as
os_event_t and wait for it directly.

Additionally, this patch removes an unreachable sleep mechanism related
to recv_sys->apply_batch_on. This variable could never be true when
entering recv_apply_hashed_log_recs, because the only places which set
it to true are:
- in MEB code. But, recv_apply_hashed_log_recs is not present in MEB
  build. So, MEB setting it to true or false, had no impact on anything
  as there is no code in MEB which would read this variable
- in the recv_apply_hashed_log_recs function itself, but this function
  is called only from the main recovery thread, and always sets it back
  to false before returning

This patch also adds exploratory debug-asserts to the recovery applier
code, to verify assumptions which would let simplify it much more.

Moreover, this patch fixes a subtle race condition, that we decrement
the numer of pages waiting for changes to be applied before we release
latch on the page, and un-io_fix it - from time to time this caused an
assertion failure when we tried to invalidate the buffer pool, because
some pages were still in non-replaceable state.

A similar race condition, also fixed by this patch, was causing the
requested BUF_FLUSH_LIST to skip over io-fixed pages, if called before
io completer io-unfixed them.

Finally, this patch corrects the assert in fil_tablespace_redo_extend,
which intended to check that BP has no pages in it, but actually was
merely checking that all pages in BP are replaceable.

Change-Id: Idfc31dc8a445b1124237e020931d0eb869a7d2ad