Bug#36128335 	mysqltest: shutdown_server on Windows is not waiting for the process to end
Symptom:
MTR tests are failing on Windows on many different tests with many different symptoms after issueing `shutdown_server` command, directly or indirectly by include files.
Most failures complain about no access to files that the killed server use, in `force-rmdir` and similar, or with port being still in use, or with files being still in use during the next server instance
startup.

Root cause:
`mysqltest` `kill_process()` is not waiting for the process to fully die. Issuing `TerminateProcess()` is not enough as it is asynchronous operation.

Fix:
After `TerminateProcess()` we call `WaitForSingleObject()` to wait for the process to fully close. A new testcase is added to try to force the above scenario.

Change-Id: Ie40e4fa7ef2567281d38ab53e1fa75479abbf1d5