Bug#34096590: Hypergraph optimizer doesn't update select_* status variables [2/3, refactor, noclose]
Initial refactoring to simplify the code.

Currently, iterators are usually created when optimize() is called on
the top-level Query_expression. Query_expression::optimize() takes a
boolean parameter that tells if iterators should be created.

This patch moves the iterator creation out of
Query_expression::optimize() and into a separate function called
Query_expression::create_iterators(). Callers that previously called
optimize() with create_iterators=true now instead call
create_iterators() immediately after optimize() has returned
successfully.

The benefits of this refactoring are:

- Query_expression::optimize() becomes simpler, as it performs a
smaller and more specific task.

- The new Query_expression::create_iterators() function can be reused
in code that creates iterators for non-top-level Query_expressions in
some specific cases (like const-table optimization) where iterators
for a sub-tree of a plan are needed. This reduces code duplication.

- Having iterator creation in a separate function allows us to move it
later in the optimization phase. The patch moves it after secondary
engine optimization, so that we don't create iterators unnecessarily
for statements that will be retried in a different engine.

Change-Id: Ie871860d02654cd798e07a48fd932dfaa6d3e483