Bug#36323066: diagnostics.sql prevents upgrades due to the presence of old terminology
In the upgrade process, the server parses the routines present in the
data being upgraded. These routines may however contain commands that
no longer exist in the current server.

Such is the case of diagnostics.sql that use the old terminology for
SHOW REPLICA STATUS, and for that reason we ignore terminology parsing
errors when the server being upgraded has a version below 8.4.

Change-Id: Iecd45483e6367b05b1f790e2db4b13b5c072302d