Bug#36313793 Use EpochContext object instead of variables [2/3]
Part of Bug36313793 ndb-log-apply-status fails to work with no
replicated changes in BI.

Instead of using separate variables to track the handling of events
within the epoch, use an object that aggregates all these. Then, using
this epoch context, a function is added to assess if the epoch is
empty in light of the variables' value that reflect what happened in
the epoch.

Also gives the possibility to have it extended.

Change-Id: I85e8b25cc3890786f97ed0f615287b3ab7e63765