Bug#36313793 --ndb-log-apply-status fails to work with no replica changes [3/3]
Problem:

The NDB table, ndb_apply_status, tracks the replica applier state. If
replicated changes were applied but are not visible in the Binlog
Injector (log-replica-updates=OFF) - because NDB data nodes filtered
those - then the epoch, from the Binlog Injector point-of-view, is
deemed to be empty.  But, ndb_apply_status did have an update because
those changes were replicated in the cluster. This breaks the flow of
apply status information in a chain/circular replication deployment.

Analysis:

Running some chain/circular replication scenarios, it is possible to
observe, in the binlog injector handling of data events, that the
ndb_apply_status table updates are written into the binlog
transaction, thus not filtered (system tables do not have
skip-log-replica-updates data node filter).

If only ndb_apply_status updates are written, then it's considered an
empty epoch, as there's no actual table data to write to the
binlog. Therefore, the binlog transaction will be rolled back, in case
logging of empty epochs is off. But, due to the nature of the
`skip-log-replica-updates` filter in the NDB data nodes, some data
might have been replicated even if not received by the binlog thread.

Solution:

The usage of ndb-log-apply-status=ON with circular replication is a
complex case. Requires the understanding that even for a server that
has skip-log-replica-updates, ndb_apply_status updates are still
logged (and thus continue to propagate through the chain).

With the above, this patch does not enable the replica updates filter
in the NDB data nodes for the cases that ndb-log-apply-status or
ndb-log-orig are ON. If log-replica-updates is OFF (filter might be
enabled and in operation) then ndb-log-apply-status and ndb-log-orig
cannot be toggled.

Tests adapted to this new behavior.

Change-Id: I48944c536142b87d361e73eacf0f7a699616fa66