BUG#36229998: Missing information on the error log about backlog apply on member join
When a member is rejoining a group, if it has transactions to apply on
`group_replication_applier` channel from its previous participation on
the group, those transactions will be applied when the member rejoins
before connections to a donor during distributed recovery.
This backlog apply can be monitored on the
`performance_schema.replication_applier_status_by_worker` table, but is
there is no information about in on the error log.
This may cause misinterpretations when reading the error log, giving the
wrong impression that the server is stalled.

Messages were added to error log to indicate start and finish of backlog
apply.

Change-Id: Ie8e64e28e7861645ce6c136906e28e570b2912b1