Bug#34595073 : MEB 8.0.30 BACKUP FAILS WHEN                PERFORMANCE_SCHEMA IS OFF
Analysis:
---------
When performance_schema is set to off, MEB fails with error
"Unexpected number of rows from MySQL query" while taking a
backup.

The tables innodb_redo_log_files & keyring_component_status
are not available when performance schema is disabled on the
server. When MEB tried to query these tables, it gets empty
set as the result and hence it fails.

Fix:
----
The fix is to make these tables available even when
performance_schema set to OFF or disabled on the server.

Change-Id: Ib3b27ebc003d533e5b843f19cf835a6caf72403f