Bug#34338001 	Performance of Temptable is worse than Memory in GROUP BY scenario
Symptom:
Some queries with `SELECT ... GROUP BY` can be a few times slower when executed on TempTable than ones executed on Memory temporary table engine.

Root cause:
The `AllocatorState::current_block` gets allocated and quickly deallocated in case a single `Row` instance is allocated on it and released in a loop.

Fix:
The `AllocatorState::current_block` is not released when it gets empty. We release it only when `AllocatorState` is deleted, that is when the `Table` gets deleted.

Additional fixes:
`AllocationScheme` gets a new `block_freed()` method to be able to be fully responsible for managing memory usage reporting to `MemoryMonitor`.
`AllocatorState` knows the `AllocationScheme` and can on its own report usage of memory to it.

Change-Id: I64ae2387dc23b3f8d027d4050972bf126aa5d004