Bug#34096590: Hypergraph optimizer doesn't update select_* status variables [3/3, fix]
Move code that collects information about the chosen query plan from
make_join_readinfo(), which is only used by the old optimizer and
which only understands the old optimizer's data structures, to a new
function, CollectStatusVariables(), which retrieves the information
from access paths and therefore works with both optimizers.

The status variables distinguish between the first table in a query
block and the other tables. The new code takes advantage of
WalkAccessPaths() visiting paths depth-first and left-to-right, so
that it visits the leftmost table of the join plan for a query block
first. The leftmost table of the join plan is accessed first during
execution, and it is read fully only once. The tables that are not
first may be read fully multiple times, at least if they are inner to
some nested loop join in the plan.

One change had to be made to WalkAccessPaths() to ensure that the
tables were seen in the desired order: It used to visit the probe
table of a hash join before the build table, but it will now visit the
build table first, to match the order in which the tables are read
during execution.

Tests that previously were disabled for the hypergraph optimizer
because it didn't update the status variables, are now enabled for
both optimizers.

Change-Id: If7216e212bfa04d0d72c2f212fff1912019449ed