BUG#36059098: Stuck group_replication_set_as_primary
It was reported that the SQL function
`group_replication_set_as_primary()` after successfully set a new
primary remained waiting for the operation completeness.

On the occurrence and implementation analysis, after ruling out
other hypothesis, the only probable cause for the issue is a
missed/not_handled broadcast on a condition wait call.
More precisely on:
```
  mysql_mutex_lock(&group_thread_end_lock);
  while (action_running && !coordinator_terminating) {
    DBUG_PRINT("sleep",
               ("Waiting for the group action execution process to terminate"));
    mysql_cond_wait(&group_thread_end_cond, &group_thread_end_lock);
  }
  mysql_mutex_unlock(&group_thread_end_lock);
```
which can cause the coordinator of the SQL function
`group_replication_set_as_primary()` to wait uninterruptedly.

To avoid the above issue, the `mysql_cond_wait()` function was
replaced by `mysql_cond_timedwait()` which will periodic check if
the predicate to wait still stands or if the wait can end.

Additionally, a superfluous call to `mysql_cond_broadcast()` was
deleted.

Change-Id: I41fa0f1b3f3ee644e19b916814837ffe5232e265