Bug#36357761 Clean up import and export of symbols for the strings library
As part of
    WL#14467: Isolate character set and collation library
we created a static and a shared version of the strings library.
The ifdefs, and cmake code, for dllexport/dllimport of
symbols is more complex than necessary.

Simplify and clean it up.

Change-Id: I9d4888cb72169d8505b39138edba2af42dd6caa4
(cherry picked from commit 38fa4d24cff8c813a3bd218b3ec4c5f0285456f1)