WL#16071 Alterring inferred data type for NDJSON
Currently guided load does not allow type casting even in non-strict mode.
So, if a json file has a number of invalid json values, the column inferred as
varchar but since varchar is not allowed for json format, the table cannot be
created and so load never initiated.
This branch solves this issue by letteing inferred data type to cast to JOSN
if the probability of invalid json values is not too high. Currently the threshold
is very conservative, i.e., we cast to JSON if only one valid JSON is seen but we
may relax this threshold in future.

Change-Id: I65af67c1251a80141880659224bb69bae19ebe2c