Bug#36118590: Inconsistent results caused by JOIN tables and CHAR func
The symptoms for this bug is that we get a different result when
a column reference argument to the CHAR function is replaced with
a CASE expression that is essentially the same as the column reference.
The CHAR function is placed in the WHERE clause of an outer join, and
the column reference is from the inner table of the outer join.
Here is the query with the CASE expression as the CHAR function argument

  SELECT 1 AS c_0
  FROM t0 LEFT JOIN t1 ON t0.vkey = t1.vkey
  WHERE CHAR(CASE WHEN FALSE THEN t1.vkey ELSE t1.vkey END) NOT LIKE 'X';

The wrong value is had when the column reference is used directly.
The reason for this is that the CHAR function in a WHERE clause is used
to convert the outer join to an inner join. However, this is only
correct when a NULL value as argument implies a NULL value as result,
which is not the case for the CHAR function. The implementation detail
that enforces this conversion is that the function's not_null_tables()
property function returns the map bit of the table.

But when the column reference is replaced with the said CASE expression,
the CASE will not propagate this not_null_tables() value, thus we
will not optimize the outer join to an inner join, and hence the result
of the operation is correct.

The fix for this problem is to set the CHAR function's null_on_null
property to false instead of true. This will ensure that
not_null_tables() returns 0 and not the table's map bit, which will
again ensure that the outer join is not converted to an inner join.

Change-Id: I8a8f5aec4df4a2bd97066f44f130cea05c20ed6a