Bug#36362495 routing plugin should not support option disabled
The routing plugin declare the "disabled" option as a supported
configuration option even though configuring it does not have any
effect. It was added to the supported options by mistake as
a copy/paste error.

This patch removes "disabled" from the list of the supported routing
configuration options. The Router will now error out if it is
configured.

Change-Id: Ia53e9d00769ade1156813965cfe7a86c6c8d25c8