WL#15742 connection-sharing without connection-pool - postfix
If the cluster was setup using hostname instead of IP-addresses, the
stashed connections weren't found.

The lookup from the stash was done via destinations-names (aka what
comes from the metadata_cache which may be hostname or IP),
but storing connections into the stash happens via endpoints which as
IPs.

Change
======

- Lookup connections in the stash by ip-address.

Change-Id: Icb560033f24affd7a500787480d535924435e29a