Bug#36324900 Ignore ENOENT error from unlink() operation
If one put the NDB data node filesystem on some distributed filesystems
one can experience that data node fails when it tries to remove a file
due to file is reported to not exist.  For a local filesystem that
should be impossible to happen, but for some distributed filesystems can
as part of internal failover handling retry a removal which may have
succeeded before the failover in which case the second removal may fail
since file no longer exists.

Data node was changed to allow file and directory removal to fail with
'file does not exist' error and treat that as successful removal.

Note, this apply only to files under the data node filesystem and backup
files. Removing other files will not have changed semantics. Nor will
any file removal by other NDB programs change semantics.

For testing purposed an extra file deletion call is issued roughly 1% of
the times a file deletion is requested, this is only done for debug
builds.

Change-Id: Ie8f5f587e9e675c2a0705d7e450be0e139b045a8