Bug#36364619: Assertion 'src != dst || cs->casedn_multiply == 1'
agg_item_set_converter() was recently updated to set the collation
for an argument item in some cases, however this can be dangerous
when the item's metadata relies upon collation data. This is e.g.
the case for the LOWER and UPPER functions, which set "multiply"
based on the collation. At runtime, the "multiply" argument is not
consistent with the actual collation, and the server gets in trouble.

The solution to this is to restrict setting collation to literal
strings only, since these are the only ones that really need it,
and does not have other metadata that depend upon the collation.

Also extend setting the collation when the desired collation is binary.
This was required due to failing range tests with the hypergraph
optimizer. It was also noticed that the collation field and
the collation of str_value inside class Item_string had to be
synchronized.

Finally, this caused problems for some statement digest tests when the
input string is binary. It was decided that this is not a reasonable
input type, so we now give an explicit error message for this case.

Change-Id: I356e010f7b5fe0a667d63b7d66bbaea1e6af8e5a