Bug#34951115: C API Hangs When Receiving Results for Server Side Prepared Statement
Problem: mysql_stmt_prepare() hanged when a parameter of the prepared
statement was used in ORDER BY clause. The server set such parameter
as hidden as it is not allowed at that place. The message sent back
to the client contained global number of parameters instead of visible
ones and data referring to only visible parameters. The function hanged
waiting for data that was not sent.

Fix:
1) Do not set hidden flag for const for execution items.
2) Assert that all parameters are visible before sending the data to the client.

Test: MySQL client test running prepared statement with parameter
in ORDER BY clause.

Change-Id: Ic2b4f9c473baa5dab062aa7e6b6a698a837f96e1