Bug#36041032 : Setting (debug only) InnoDB system variables                causes crash on Debug Server build
Description:
------------
The Server crashes when setting "innodb_interpreter_output"
variable.

Analysis:
---------
The "innodb_interpreter" and "innodb_interpreter_output"
variables are meant to be debug only variables that are
added for internal testing purposes. "innodb_interpreter"
variable can be used to execute commands and its output can
be accessed via "innodb_interpreter_output" variable. As,
the latter variable is used to store the result of the
commands executed, it should not be allowed to be set by
the user.

Fix:
----
Make the "innodb_interpreter_output" variable as a read only
variable.

Change-Id: Ice3883975727d9bad9322ca9c46458973a9b61ba