Bug#36361165 Incorrect query result caused by comparing decimal and float type value
Also solves duplicate Bug#36492878.

In analyze_decimal_field_constant, we have a place where we truncate
the constant towards zero to fit the same number of decimals as the
column/field:

    if (decimal_round(&cpy, &cpy, f_frac, cpy.sign() ? CEILING : FLOOR))

Just after that we make a note of whether this constant was a negative constant:

    *negative = cpy.sign();

The problem is that in this corner case, where we shed all decimals
and truncate to exactly zero, cpy.sign() will no longer return true
*after* the rounding. This causes later logic of adjusting operators
to malfunction.

Solution: make note of this fact *before* we truncate towards zero.
Change-Id: Ide9bd441925bb3fb5f11595a9f53bdbf61cc6acc