Bug#36086735 Cluster/J tests report error in my_thread_global_end()
The NDB client cannot shut down cleanly unless the Cluster/J test
suite closes the main SessionFactory. This patch creates a
ClearSmokeTest in the test suite, responsible for closing the
SessionFactory, and adapts the test runner to always run the
ClearSmokeTest at the conclusion of the suite.

With this patch, the number of threads reported as not exiting
should be reduced to 1. The final thread is an unrelated issue,
36383937, and will be fixed separately.

This patch also changes testClusterJ to fail quickly whenever
there is no possible path to the Connector/J JDBC driver.

Change-Id: I266a02d7f6eaa1d71436d0b055e5db575d66788b