Bug #35835864 : Crash during background rollback if both prepare and active transaction exist
SYMPTOM:
-------
Assertion failure in InnoDB's background
when a transaction for which it wants to acquire an MDL lock, turns out
to be no longer active.

ROOT CAUSE:
----------
When creating a list of <trx id, table id> pairs for which we will need
to acquire MDL in trx_recovery_rollback_thread later, we accidentally
include a transaction which not only will not be rolled back by InnoDB
background thread (because it is TRX_STATE_PREPARED), but worse still,
might be rolled back by  even before we get to
spawn .

SOLUTION:
--------
InnoDB should never include tables of transactions which are in
TRX_STATE_PREPARED state in the list, because InnoDB is not allowed to
roll them back by itself, and they might be rolled back by binlog logic.

Thanks to Genze Wu (Alibaba) for the contribution.

Change-Id: I9f366c1e1022464e6dd43de08bed89f0510ad786