Bug#36339892 ClusterSet.listRouters() reports wrong version after upgrade
When the Router is bootstrapped to the fresh directory it tries to
create a  new account in the `routers` metadata table. This table has
a unique constraint UNIQUE KEY `address` (`address`,`router_name`).
If there already is a record with that address and router_name pair
the insert fails due to the contraint mentioned before. If no --force
parameter is given the Router fails to bootstrap with an error
as expected.
However if the --force parameter is provided and there already is
a record matching (`address`,`router_name`) pair, the Router queries for
the router_id to re-use it. The problem is that the query that does it
is invalid, it only has "WHERE router_name=?" clause, ignoring the
address part. This means that it can return multiple rows for any
router with a given name for various hosts/addresses. Then the id of the
random (first from the list of the rows) is taken. This can lead to
multiple Routers on various host using the same account, updating
the attibutes etc.

This patch fixes the query retrieving the router_id so that it matches
the unique constrain (address, router_name).

Change-Id: I8f8f7227e9597cbd5a917d8ef068c7b6214c5288