Bug#36259487 	InnoDB: OS_FILE_NORMAL and OS_FILE_AIO are very misleading
Symptom:
The use of `OS_FILE_NORMAL` leads to hard to debug performance regressions on Windows.

Root cause:
The `os_file_create_func()` function has an argument `purpose` that has two valid values: `OS_FILE_NORMAL` and `OS_FILE_AIO`. They are not documented and don't do what you would expect them to do.
This easily leads to confusion and very hard to pinpoint bugs. For example bug fixed with `Bug#36154818 	[InnoDB] Redolog files on Windows are not using Overlapped mode` fixed a huge regression.
The problem is that `OS_FILE_AIO` is in general the "normal" mode, and the `OS_FILE_NORMAL` should be used only if it is well documented and tested to be giving any advantage.
In general: this is Windows-only option. It does nothing for any other system.
`OS_FILE_AIO` opens file in Overlapped mode (with `FILE_FLAG_OVERLAPPED`). This causes most if not all calls to WriteFile and ReadFile to return FALSE and `GetLastError()` to return `ERROR_IO_PENDING` 997
"Overlapped I/O operation is in progress.". The user needs to call `GetOverlappedResult()` once the event passed in the `OVERLAPPED` structure gets signaled. On the other hand, if this flag is not specified on
CreateFile, then all file accesses are serialized in kernel, including IOs like `WriteFile`, `ReadFile`, but also `fsync()`, that is `FlushFileBuffers()`. This is a direct cause of the regression in the
Bug#36154818.
Historically the Overlapped mode was used for InnoDB AIO. The control flow was transferred to a io_handler thread that was waiting for a signaled event and was post-processing the finished IO.
A file opened with OS_FILE_AIO mode could only issue IOs with `os_aio()`. If need be, AIO_SYNC mode for `os_aio()` was used to make the call a blocking one.
A file opened in OS_FILE_NORMAL mode could only use `os_file_write()`/`os_file_read()`.
However, some time ago in `Bug #33840645 	InnoDB: AIO doesn't need SYNC queue` the logic of `os_file_write()` was altered - we no longer care if file was opened in OS_FILE_NORMAL or OS_FILE_AIO mode, both were
working correctly. This made the `AIO_SYNC` mode obsolete and was just removed. At this time the OS_FILE_AIO mode lost its direct connection with the InnoDB AIO subsystem.
The `OS_FILE_AIO` lost its different usage requirements. This caused the `OS_FILE_NORMAL` to virtually become "the same as OS_FILE_AIO, but all accesses to the file are serialized".
This poses a question: why would anyone want to use OS_FILE_NORMAL mode then, if it is just worse?
It turns out that the non-Overlapped file has one advantage: If you read or write bigger buffers from exactly one thread, especially in linear fashion, it may lead to 0-5% better throughput/performance than
with Overlapped mode.
So, basically to use 'OS_FILE_NORMAL` one should document exactly that: exactly one thread is using the file at a given time, and using it brings real performance gain.
This task is to rename and document these flags appropriately.

Fix:
The `ulint purpose` is removed from signature of `os_file_create()`.
The `OS_FILE_AIO` and `OS_FILE_NORMAL` are removed.
The `ulint type` of `os_file_create()` method is renamed to `purpose` (after the previous `purpose` argument was removed as described a few lines above).
A new `OS_LOG_FILE_RESIZING` file opening purpose for opening log files for short time, when they are being filled with zeros.
On Windows the `os_file_create()` uses `FILE_FLAG_OVERLAPPED` attribute, but disables it only for `type` being `OS_CLONE_LOG_FILE`, `OS_CLONE_DATA_FILE`, or `OS_LOG_FILE_RESIZING`.
This basically allows for the specified list of the file types to be opened in the serialized IO mode, and one can add new elements to this list, but should:
- not use any of the existing types, as these are to be used for concurrent access
- prove it brings real gain in testing and it is used really only for serialized IO.

Additional fixes:
Double-write buffer files are not opened for serialized IO anymore. It gives up to 60% higher performance on Windows on some tests.
Undo files are not opened for serialized IO anymore in `srv_start.cc`.
`WIN_ASYNC_IO` is removed and replaced by just `_WIN32`. This was used to distinguish Windows 95-based systems, which are long gone.

Change-Id: I5acf4ecb06641de0b48429a0ed3a2189b4c4a06f